re-organize so that tensorflow etc are truly optional
now can use non-tensorflow classes without having it in the environment
