+++ requirements.txt
@@ -1,9 +1,9 @@
 numpy
 pandas
 torch
-tensorflow_hub
-tensorflow
 tqdm
 opensoundscape
-timm
-torchaudio
+# timm
+# torchaudio
+# tensorflow_hub
+# tensorflow
